update comment function in my-web.

+++ src/main/java/net/slipp/web/QuestionController.java
@@ -1,5 +1,6 @@
 package net.slipp.web;
 
+//import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -10,6 +11,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+
 import net.slipp.domain.Question;
 import net.slipp.domain.QuestionRepository;
 import net.slipp.domain.User;
@@ -42,28 +44,56 @@ public class QuestionController {
 	
 	@GetMapping("/{id}")
 	public String readQuestion(@PathVariable long id, Model model) {
-		Question question = questionRepository.findOne(id);
-		model.addAttribute("question", question);
+		model.addAttribute("question", questionRepository.findOne(id));
 		return "/qna/show";
 	}
 	
 	@GetMapping("/{id}/form")
-	public String updateForm(@PathVariable Long id, Model model) {
-		model.addAttribute("question", questionRepository.findOne(id));
-		System.out.println(questionRepository.findOne(id));
+	public String updateForm(@PathVariable Long id, Model model, HttpSession session) {
+		if (!HttpSessionUtils.isLoginUser(session)) {
+			return "redirect:/user/loginForm";		//login 되어있는지 확인. qnaSubmit() 에서 복, 붙
+		}
+		
+		User loginUser = HttpSessionUtils.getUserFromSession(session);
+		Question question = questionRepository.findOne(id);
+		if (!question.isSameWriter(loginUser)) {
+			return "redirect:/user/loginForm";
+		}
+		
+		model.addAttribute("question", question);
+		System.out.println(question);
 		return "/qna/updateForm";
 	}
 	
 	@PutMapping("/{id}")
-	public String update(@PathVariable Long id, String title, String contents) {
+	public String update(@PathVariable Long id, String title, String contents, HttpSession session) {
+		if (!HttpSessionUtils.isLoginUser(session)) {
+			return "redirect:/user/loginForm";		//login 되어있는지 확인. qnaSubmit() 에서 복, 붙
+		}
+		
+		User loginUser = HttpSessionUtils.getUserFromSession(session);
 		Question question = questionRepository.findOne(id);
+		if (!question.isSameWriter(loginUser)) {
+			return "redirect:/user/loginForm";
+		}
+		
 		question.update(title, contents);
 		questionRepository.save(question);
 		return String.format ("redirect:/questions/%d", id);
 	}
 	
 	@DeleteMapping("/{id}")
-	public String delete(@PathVariable Long id) {
+	public String delete(@PathVariable Long id, HttpSession session) {
+		if (!HttpSessionUtils.isLoginUser(session)) {
+			return "redirect:/user/loginForm";		//login 되어있는지 확인. qnaSubmit() 에서 복, 붙
+		}
+		
+		User loginUser = HttpSessionUtils.getUserFromSession(session);
+		Question question = questionRepository.findOne(id);
+		if (!question.isSameWriter(loginUser)) {
+			return "redirect:/user/loginForm";
+		}
+		
 		questionRepository.delete(id);
 		return "redirect:/";
 	}
